+++ src/polyhedra/include/exaDEM/compute_vertices.hpp
@@ -26,7 +26,7 @@ namespace exaDEM
   struct PolyhedraComputeVerticesFunctor
   {
     shapes& shps;
-    ONIKA_HOST_DEVICE_FUNC inline void operator () (const uint8_t type, const double rx, const double ry, const double rz, const double h, const exanb::Quaternion& orient, ::onika::oarray_t<::exanb::Vec3d, EXADEM_MAX_VERTICES>& vertices ) const
+    ONIKA_HOST_DEVICE_FUNC inline void operator () (const uint32_t type, const double rx, const double ry, const double rz, const double h, const exanb::Quaternion& orient, ::onika::oarray_t<::exanb::Vec3d, EXADEM_MAX_VERTICES>& vertices ) const
     {
 			// h will be used in a next development
 			const auto& shp = shps[type];
